context_engineering/pyproject.toml
[project]
name = "context-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34.0",
    "colorama>=0.4.6",
    "rich>=13.7.0",
    "strands-agents[openai]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.2.2",
]
